chore: parameterize trade input header top-right content
